Establish a fakemozilla Puppet class
We want to allow parties without access to Mozilla's private Puppet
repository to recreate the Mercurial server environment as closely as
possible. To facilitate this, we introduce the "fakemozilla" class
for performing base system configuration in a way that approximates
Mozilla's systems. It isn't perfect, but it is better than none.